requirements.txt
av==10.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.3
coloredlogs==15.0.1
ctranslate2==3.12.0
faster-whisper==0.4.1
ffmpeg-python==0.2.0
filelock==3.12.0
flatbuffers==23.3.3
future==0.18.3
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
lit==16.0.1
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.14.1
openai-whisper @ git+https://github.com/openai/whisper.git@c09a7ae299c4c34c5839a76380ae407e7d785914
packaging==23.1
protobuf==4.22.3
PyAudio==0.2.13
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
SpeechRecognition==3.10.0
sympy==1.11.1
tiktoken==0.3.1
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
